chore: clear pre-existing engine lint, format, mypy, and test failures
Part 2 of the report-formats sweep. With the engine feature work
already shipped on main, this commit lands the cleanup that had been
blocking a clean pre-commit run on any engine change:

ruff (was: 72 issues)
- TC001/TC002/TC003: moved runtime-only imports into TYPE_CHECKING
  blocks across imports/, reports/, api/routes/
- I001: reordered import blocks in admin.py and friends
- RUF012: annotated class-level mutable defaults (acrobat/pitstop
  parsers) with `ClassVar[set[str]]`
- RUF059: prefixed unused tuple-unpack vars with `_` in viewer.py
- B904: added `from None` to HTTPException re-raises in
  approvals.py (6 spots) so the 404s no longer chain through the
  suppressed ValueError
- B007, SIM105, SIM108, SIM113, F841, F401, UP037: small local
  cleanups applied via `ruff check --fix` and hand-review
- 42 files reformatted to match the repo's ruff-format config

mypy (was: 15 errors)
- Added `plugins = ["pydantic.mypy"]` to pyproject so Pydantic
  BaseModel subclasses stop reading as `Any` (11 errors gone)
- approvals/schemas.py: parameterised `list[dict]` -> `list[dict[str, Any]]`
- integrations/railway.py: assigned `response.json()` to a typed
  local so the `no-any-return` goes away
- reports/service.py: added `-> Any` return annotation to the
  `_bytes_stream` helper
- reports/separation_renderer.py: widened `tac_limit: int = 300`
  to `tac_limit: float = 300` so the viewer's float TAC limit no
  longer fails `arg-type`
- api/routes/jobs.py: narrowed the `external_report.content_type`
  access to handle the `UploadFile | None` shape the union-attr
  checker was flagging

pytest (was: 7 failures on baseline)
- tests/test_reports_engine.py: updated
  `test_dispatches_to_html|pdf` and `test_delegates_to_html|pdf_report`
  mocks to assert on the full `(result, branding=None,
  pdf_bytes=None, detail_level="standard")` call signature that
  `ReportEngine.generate` has used since the branding work landed
- tests/ai/test_orchestrator_ai.py: the orchestrator appends an
  `AI_SCAN_001` audit-trail marker whenever any AI analyzer runs.
  Updated `test_ai_enabled_runs_analyzers`,
  `test_ai_findings_merged_with_engine_findings`, and
  `test_ai_findings_have_correct_source_and_category` to filter
  the marker out (`category != "ai_scan"`) and to expect
  `ai_findings_count == 2` (1 real + 1 marker) where the tests
  were asserting pre-marker values
- reports/templates/report.html: added an explicit "No findings
  detected." empty state after the finding list so the
  `test_empty_result_shows_no_findings` contract holds again

Net: `ruff check src/ tests/` clean, `ruff format --check` clean,
`mypy src/` clean (240 files), `pytest tests/` clean (1814 passed,
0 failed — excluding the `tests/e2e/` suite which needs a live
database).

+++ packages/engine/src/lintpdf/reports/separation_renderer.py
@@ -248,7 +248,7 @@ def render_tac_heatmap(
     pdf_bytes: bytes,
     page_num: int,
     dpi: int = 150,
-    tac_limit: int = 300,
+    tac_limit: float = 300,
 ) -> bytes:
     """Generate a TAC (Total Area Coverage) heatmap PNG overlay.
